Moved stun mags/ammo
