--- BNA_KC_Weapons/12thMags.hpp
@@ -1,20 +1,0 @@
-// Stun
-class JLTS_stun_mag_long;
-class 12thFleet_Mag_StunLong: JLTS_stun_mag_long
-{
-    displayName = "[12th Fleet] Stun Energy Cell (Long)";
-    displayNameShort = "Stun Rounds";
-    descriptionShort = "Stun Energy Cell <br />Rounds: 5 <br />Duration: 15 Seconds <br />Used in: DC-15S, DC-15A, DC-17";
-
-    ammo = "12thFleet_Ammo_Stun";
-};
-
-class JLTS_stun_mag_short;
-class 12thFleet_Mag_StunShort: JLTS_stun_mag_short
-{
-    displayName = "[12th Fleet] Stun Energy Cell (Short)";
-    displayNameShort = "Stun Rounds";
-    descriptionShort = "Stun Energy Cell <br />Rounds: 10 <br />Duration: 5 Seconds <br />Used in: DC-15S, DC-15A, DC-17";
-
-    ammo = "12thFleet_Ammo_Stun";
-};
